requirements.txt
fastapi
uvicorn
twilio
gspread
oauth2client
python-dotenv
httpx
